+++ pyproject.toml
@@ -1,0 +1,34 @@
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "chenla_diffusion2D"
+authors = [
+    {name = "LingChia Chen", email = "st191429@stud.uni-stuttgart.de"},
+]
+description = "This is a package to show plots of the solutions of a diffusion2D problem."
+readme = "README.md"
+requires-python = ">=3.6"
+keywords = ["diffusion2D", "plots"]
+license = {text = "BSD-3-Clause"}
+classifiers = [
+    "License :: OSI Approved :: BSD License",
+    "Development Status :: 2 - Pre-Alpha",
+    "Programming Language :: Python :: 3",
+]
+dependencies = [
+    "numpy",
+    "matplotlib",
+]
+version = "0.0.3"
+
+[project.urls]
+Homepage = "https://github.com/lingachen/diffusion2D/tree/chenla-dev"
+
+[project.optional-dependencies]
+pdf = ["ReportLab>=1.2", "RXP"]
+rest = ["docutils>=0.3", "pack ==1.1, ==1.3"]
+
+[project.scripts]
+chenla_diffusion2d = "chenla_diffusion2D.diffusion2d:solve"
